Dragging And Dropping!
Dragging and dropping now work; I made a lot of changes to the dragging part that worked before, though. The way I was doing things (catching events in the tiles themselves), I wasn't able to easily to get dropping to work without adding global state (ugh). So I reworked the handlers to fire in the tile container instead; that way, everything I needed would bubble up and I could handle it based on the event target.